pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["vadimo"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"

[tool.poetry.group.dev.dependencies]
cowsay = "^5.0"
flake8 = "^4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
